Added components group for app utilities and extensions.
