+++ Kwartracker/Custom Components/View/View+Extensions.swift
@@ -1,0 +1,14 @@
+//
+//  View+Extensions.swift
+//  Kwartracker
+//
+//  Created by John Ellie Go on 4/15/21.
+//
+
+import SwiftUI
+
+extension View {
+    func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
+        clipShape( RoundedCorner(radius: radius, corners: corners) )
+    }
+}
